Move arithmetic functions into submodule `FixedPointArithmetic`
